Remove all non-TC messages from console output
Thanks to @sergiorykov for pointing out in Issue #2 that this reporter
shouldn't send anything other than teamcity messages to the console.
Most of these messages were added while debugging the teamcity messages
and are no longer necessary.

I also fixed a bug I noticed while testing the change: if a test does
not pass and does not have a response property the reporter will crash
because we were accessing `code` and `reason` without checking
that `this.currItem.response` exists.