Switch from expect / mocha to jest
The main thing I want to test is the output to console.log of the
reporter. Because expect / mocha don't really provide this functionality and
expect has actually been "donated" to jest I decided to switch to jest.

I also added a test that verifies that the messages spat out by the
events (when triggered with basically no arguments) are all valid TC
messages. This test could be more thorough in its validation of the
message format. We could also add to the arguments so that we test the
main branches in the event handlers.